Moved discussion service into lms runtime
